Refactor capture pane components
