+++ src/renderer/components/animator/ImagingDeviceSettings/ImagingDeviceSettingName/ImagingDeviceSettingName.tsx
