+++ src/renderer/components/animator/ImagingDeviceSettings/ImagingDeviceSettingsTr/ImagingDeviceSettingsTr.tsx
