+++ src/renderer/components/animator/ImagingDeviceSettings/ImagingDeviceSettingsTrBoolean/ImagingDeviceSettingsTrBoolean.tsx
