+++ src/renderer/components/animator/ImagingDeviceSettings/ImagingDeviceSettingsTrList/ImagingDeviceSettingsTrList.tsx
