+++ src/renderer/components/animator/ImagingDeviceSettings/ImagingDeviceSettingsTrRange/ImagingDeviceSettingsTrRange.tsx
