--- src/renderer/components/animator/ImagingDeviceSettingsTable/ImagingDeviceSettingsTable.css
@@ -1,14 +1,0 @@
-.settings-table-row td:nth-child(1) {
-  max-width: 6rem;
-  text-wrap: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  line-height: var(--mantine-line-height-xs);
-}
-
-.settings-table-row td:nth-child(2) {
-  min-width: 6rem;
-}
-.settings-table-row td:nth-child(3) {
-  width: 5rem;
-}
